Creating a fail-safe for the user chosen values... Added an option for a limiter

+++ Simple Sound Generator-CIS 1202 Final (Project)/Sound.h
@@ -37,7 +37,7 @@ class Sine {
 public:
 	//setup for the sine sound
 	const int maxAMPLITUDE = (pow(16, 2) / 2) - 1;//the number of bits to the second power divided by 2 (to factor both negative and positive values) while subtracting 1 to factor in zero
-	int duration;//how long the sound will be (user will input the duration in seconds)
+	float duration;//how long the sound will be (user will input the duration in seconds)
 	double frequencyInput;
 
 };
